Group polynomialization and quadratization; Fix output of ID-DLS
