+++ qbee/__init__.py
@@ -1,4 +1,4 @@
 from .structures import EquationSystem, derivatives
 from .polynomialization import polynomialize
-from .algorithm import BranchAndBound, ID_DLS
+from .quadratization import BranchAndBound, ID_DLS
 from .vizualization import visualize
